Network.Google.Auth

Description

Explicitly specify your Google credentials, or retrieve them from the underlying OS.

Credentials

data Credentials (s :: [Symbol])

The supported credential mechanisms.

Constructors

FromMetadata !ServiceId
Obtain and refresh access tokens from the underlying GCE host metadata at http://169.254.169.254.

FromClient !OAuthClient !(OAuthCode s)
Obtain and refresh access tokens using the specified client secret and authorization code obtained from.

See the OAuth2 Installed Application documentation for more information.

FromAccount !ServiceAccount
Use the specified service_account and scopes to sign and request an access token. The ServiceAccount will also be used for subsequent token refreshes.

A ServiceAccount is typically generated through the Google Developer Console.

FromUser !AuthorizedUser
Use the specified authorized_user to obtain and refresh access tokens.

An AuthorizedUser is typically created by the gcloud init command of the Google CloudSDK Tools.

Application Default Credentials

getApplicationDefault :: (MonadIO m, MonadCatch m) => Manager -> m (Credentials s)

Performs credentials discovery in the following order:

Read the default credentials from a file specified by the environment variable GOOGLE_APPLICATION_CREDENTIALS if it exists.

Read the platform equivalent of ~/.config/gcloud/application_default_credentials.json if it exists. The ~/.config component of the path can be overriden by the environment variable CLOUDSDK_CONFIG if it exists.

Retrieve the default service account application credentials if running on GCE. The environment variable NO_GCE_CHECK can be used to skip this check if set to a truthy value such as 1 or true.

The specified Scopes are used to authorize any service_account that is found with the appropriate OAuth2 scopes, otherwise they are not used. See the top-level module of each individual gogol-* library for a list of available scopes, such as Network.Google.Compute.computeScope.

See: Application Default Credentials

fromWellKnownPath :: (MonadIO m, MonadCatch m) => m (Credentials s)

Attempt to load either a service_account or authorized_user formatted file to obtain the credentials neccessary to perform a token refresh.

The specified Scopes are used to authorize any service_account that is found with the appropriate scopes, otherwise they are not used. See the top-level module of each individual gogol-* library for a list of available scopes, such as Network.Google.Compute.computeScope.

See: cloudSDKConfigPath, defaultCredentialsPath.

fromFilePath :: (MonadIO m, MonadCatch m) => FilePath -> m (Credentials s)

Attempt to load either a service_account or authorized_user formatted file to obtain the credentials neccessary to perform a token refresh from the specified file.

The specified Scopes are used to authorize any service_account that is found with the appropriate scopes, otherwise they are not used. See the top-level module of each individual gogol-* library for a list of available scopes, such as Network.Google.Compute.computeScope.

saveAuthorizedUserToWellKnownPath

Arguments

:: (MonadIO m, MonadCatch m)

=> Bool
Force to save if True

-> AuthorizedUser

-> m ()

Save AuthorizedUser See: cloudSDKConfigPath, defaultCredentialsPath.

saveAuthorizedUser

Arguments

:: (MonadIO m, MonadCatch m)

=> FilePath

-> Bool
Force to save if True

-> AuthorizedUser

-> m ()

Save AuthorizedUser

Service account user impersonation

serviceAccountUser :: forall s. AllowScopes s => Maybe Text -> Credentials s -> Credentials s

Set the user to be impersonated for a service account with domain wide delegation. See https://developers.google.com/identity/protocols/OAuth2ServiceAccount

Installed Application Credentials

installedApplication :: OAuthClient -> OAuthCode s -> Credentials s

Create new Installed Application credentials.

Since it is intended that the user opens the URL generated by formURL in a browser and the resulting OAuthCode is then received out-of-band, you must ensure that the scopes passed to formURL and the type of OAuthCode correctly match, otherwise an authorization error will occur.

For example, doing this via getLine and copy-paste:

{-# LANGUAGE ScopedTypeVariables #-}

import Data.Proxy     (Proxy (..))
import Data.Text      as T
import Data.Text.IO   as T
import System.Exit    (exitFailure)
import System.Info    (os)
import System.Process (rawSystem)

redirectPrompt :: AllowScopes (s :: [Symbol]) => OAuthClient -> proxy s -> IO (OAuthCode s)
redirectPrompt c p = do
  let url = formURL c p
  T.putStrLn $ "Opening URL " `T.append` url
  _ <- case os of
    "darwin" -> rawSystem "open"     [unpack url]
    "linux"  -> rawSystem "xdg-open" [unpack url]
    _        -> T.putStrLn "Unsupported OS" >> exitFailure
  T.putStrLn "Please input the authorisation code: "
  OAuthCode <$> T.getLine

This ensures the scopes passed to formURL and the type of OAuthCode s are correct.

formURL :: AllowScopes (s :: [Symbol]) => OAuthClient -> proxy s -> Text

Given an OAuthClient and a list of scopes to authorize, construct a URL that can be used to obtain the OAuthCode.

See: Forming the URL.

Authorizing Requests

authorize :: (MonadIO m, MonadCatch m, AllowScopes s) => Request -> Store s -> Logger -> Manager -> m Request

Apply the (by way of possible token refresh) a bearer token to the authentication header of a request.

Thread-safe Storage

data Store (s :: [Symbol])

Data store which ensures thread-safe access of credentials.

initStore :: (MonadIO m, MonadCatch m, AllowScopes s) => Credentials s -> Logger -> Manager -> m (Store s)

Construct storage containing the credentials which have not yet been exchanged or refreshed.

retrieveAuthFromStore :: (MonadIO m, MonadCatch m, AllowScopes s) => Store s -> m (Auth s)

Retrieve auth from storage

data Auth (s :: [Symbol])

An OAuthToken that can potentially be expired, with the original credentials that can be used to perform a refresh.

Constructors

Auth

Fields

_credentials :: !(Credentials s)

_token :: !(OAuthToken s)

authToAuthorizedUser :: AllowScopes s => Auth s -> Either Text AuthorizedUser

authToAuthorizedUser converts Auth into an AuthorizedUser by returning Right if there is a FromClient-constructed Credentials and a refreshed token; otherwise, returning Left with error message.

exchange :: forall m s. (MonadIO m, MonadCatch m, AllowScopes s) => Credentials s -> Logger -> Manager -> m (Auth s)

Perform the initial credentials exchange to obtain a valid OAuthToken suitable for authorizing requests.

refresh :: forall m s. (MonadIO m, MonadCatch m, AllowScopes s) => Auth s -> Logger -> Manager -> m (Auth s)

Refresh an existing OAuthToken.

Default Constants

checkGCEVar :: String

The NO_GCE_CHECK environment variable.

cloudSDKConfigDir :: String

The environment variable name which is used to specify the directory containing the application_default_credentials.json generated by gcloud init.

defaultCredentialsFile :: String

The environment variable pointing the file with local Application Default Credentials.

Handling Errors

class AsAuthError a where

Minimal complete definition

_AuthError

Methods

_AuthError :: Prism' a AuthError

A general authentication error.

_RetrievalError :: Prism' a HttpException

An error occured while communicating over HTTP with either then local metadata or remote accounts.google.com endpoints.

_MissingFileError :: Prism' a FilePath

The specified default credentials file could not be found.

_InvalidFileError :: Prism' a (FilePath, Text)

An error occured parsing the default credentials file.

_TokenRefreshError :: Prism' a (Status, Text, Maybe Text)

An error occured when attempting to refresh a token.

data AuthError

An error thrown when attempting to readwrite AuthNAuthZ information.

Constructors

RetrievalError HttpException

MissingFileError FilePath

InvalidFileError FilePath Text

TokenRefreshError Status Text (Maybe Text)

FileExistError FilePath

OAuth Types

data OAuthClient

A client identifier and accompanying secret used to obtain/refresh a token.

Constructors

OAuthClient

Fields

_clientId :: !ClientId

_clientSecret :: !Secret

data OAuthToken (s :: [Symbol])

An OAuth bearer type token of the following form:

{
  \"token_type\": \"Bearer\",
  \"access_token\": \"eyJhbGci...\",
  \"refresh_token\": \"1/B3gq9K...\",
  \"expires_in\": 3600,
  ...
}

The _tokenAccess field will be inserted verbatim into the Authorization: Bearer ... header for all HTTP requests.

Constructors

OAuthToken

Fields

_tokenAccess :: !AccessToken

_tokenRefresh :: !(Maybe RefreshToken)

_tokenExpiry :: !UTCTime

newtype OAuthCode (s :: [Symbol])

An OAuth client authorization code.

Constructors

OAuthCode Text

newtype OAuthScope

An OAuth2 scope.

Constructors

OAuthScope Text

Re-exported Types

newtype AccessToken

An OAuth2 access token.

Constructors

AccessToken Text

newtype RefreshToken

An OAuth2 refresh token.

Constructors

RefreshToken Text

newtype Secret

An opaque client secret.

Constructors

Secret Text

newtype ServiceId

A service identifier.

Constructors

ServiceId Text

newtype ClientId

A client identifier.

Constructors

ClientId Text

Re-exported Modules

module Network.Google.Auth.Scope